+++ src/server.ts
@@ -1,11 +1,6 @@
 import * as Hapi from 'hapi'
-import { createToken, verifyToken } from './helpers/auth'
-import { encrypt, compare } from './helpers/crypto'
-import { Document } from './services/storage/repository'
-import userRepository from './services/storage/repositories/user'
-import trackRepository from './services/storage/repositories/track'
-import { User } from './types'
 import auth from './handlers/auth'
+import user from './handlers/user'
 
 const createServer = (port: number) => {
   const server = new Hapi.Server()
@@ -23,25 +18,10 @@ const createServer = (port: number) => {
     handler: auth.login,
   })
 
-  const sanitizeUser = (user: User) => {
-    return Object.assign({}, user, { password: undefined })
-  }
-
-  /**
-   * Who am I?
-   */
   server.route({
     method: 'GET',
     path: '/me',
-    handler: (request, reply) => {
-      const token = verifyToken(request.headers['authorization'])
-      if (!token) {
-        return reply({}).code(400)
-      }
-      userRepository.find(token.userId)
-        .then(user => reply(sanitizeUser(user._source)))
-        .catch(err => reply(err).code(400))
-    }
+    handler: user.me,
   })
 
   return server
